requirements.txt
altgraph==0.17.5
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
mysql-connector-python==9.6.0
packaging==26.1
pefile==2024.8.26
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
requests==2.33.1
schedule==1.2.2
setuptools==82.0.1
urllib3==2.6.3
openpyxl==3.1.5
flask==3.1.3
flask-login==0.6.3
livereload==2.7.1
types-requests==2.33.0.20260408
